auth_backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1